agent/requirements.txt
websockets==12.0
Pillow==10.3.0
pynput==1.7.6
